Fix P0 sync and database issues
- Migrate Show ID to Long to prevent collisions
- Fix infinite sync loop in ViewModel
- Prevent data loss during Google Sign-In merge
- Convert FirebaseSyncManager to a singleton

+++ app/src/main/java/com/thisisnotajoke/marqueepass/data/ShowDao.kt
@@ -17,7 +17,7 @@ interface ShowDao {
     fun getShowsByStatus(status: ShowStatus): Flow<List<Show>>
 
     @Query("SELECT * FROM shows WHERE id = :id")
-    suspend fun getShowById(id: Int): Show?
+    suspend fun getShowById(id: Long): Show?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertShow(show: Show): Long
